Reimplement '?' opcode
; Change mod type to uint8_t;
;
; Add test.scaf program;
;
; curr_head_ptr has a bigger role now.
; The code speaks for itself I think;

+++ Head.h
@@ -12,7 +12,7 @@
 typedef struct {
     uint32_t pos_x;
     uint32_t pos_y;
-    int32_t mod;
+    uint8_t mod;
     bool destructive;
 } Head;
